requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
dataclasses==0.6
defusedxml==0.5.0
Django==2.0
django-cors-headers==2.4.0
django-extensions==2.1.6
django-extra-fields==1.0.0
django-filter==2.0.0
django-filters==0.2.1
django-rest-framework-mongoengine==3.3.1
django-smtp-ssl==1.0
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-stubs==0.1.0
djongo==1.2.23
future==0.17.1
idna==2.8
kombu==4.2.2.post1
Markdown==3.0.1
mongoengine==0.17.0
mysqlclient==1.3.14
oauthlib==2.1.0
Pillow==5.3.0
PyJWT==1.7.1
pymongo==3.8.0
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.2.4
urllib3==1.24.1
uWSGI==2.0.18
vine==1.1.4
virtualenv==16.4.3
